Rename helpers to slack_helpers
