+++ slack_main.py
@@ -3,7 +3,7 @@ from slack_bolt.adapter.socket_mode import SocketModeHandler
 from config import config
 import re
 
-from helpers.helper_functions import (
+from slack_helpers.helper_functions import (
     handle_help,
     handle_create_rosa_cluster,
     handle_create_openstack_vm,
